# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Coach Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{teamId}/members:
    get:
      operationId: listTeamMembers
      tags: [members]
      summary: List staff members of a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [members]
      summary: Invite a staff member by email (owner only)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /teams/{teamId}/members/{memberId}:
    patch:
      operationId: updateTeamMember
      tags: [members]
      summary: Change a member's role (owner only)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: removeTeamMember
      tags: [members]
      summary: Remove a member or cancel an invite (owner only)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # Shared staff notes
  /teams/{teamId}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List shared notes (pinned first, newest first)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a shared note
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /teams/{teamId}/notes/{noteId}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note (author or owner)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note (author or owner)
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # In-app notifications
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark one notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppNotification"
components:
  schemas:
...
    TeamMember:
      type: object
      required: [id, teamId, role, status, createdAt]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        userId:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, coach, assistant, analyst]
        status:
          type: string
          enum: [pending, active]
        createdAt:
          type: string
    TeamMemberInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          minLength: 3
        role:
          type: string
          enum: [coach, assistant, analyst]
        displayName:
          type: string
    TeamMemberUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [coach, assistant, analyst]
    Note:
      type: object
      required: [id, teamId, authorUserId, authorName, content, pinned, createdAt, updatedAt]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        authorUserId:
          type: string
        authorName:
          type: string
        title:
          type: ["string", "null"]
        content:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [content]
      properties:
        title:
          type: string
        content:
          type: string
          minLength: 1
        pinned:
          type: boolean
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
          minLength: 1
        pinned:
          type: boolean
    AppNotification:
      type: object
      required: [id, type, read, createdAt]
      properties:
        id:
          type: integer
        teamId:
          type: ["integer", "null"]
        type:
          type: string
        meta:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        read:
          type: boolean
        createdAt:
          type: string
...