# lib/api-zod/src/generated/types/teamMemberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Coach Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberUpdateRole } from './teamMemberUpdateRole';

export interface TeamMemberUpdate {
  role: TeamMemberUpdateRole;
}